backend/requirements.txt
fastapi
uvicorn
passlib
python-jose
python-multipart
sqlalchemy
pydantic_settings
psycopg2-binary